[CI]: Attempt building on Windows & MS-DOS
When running CI actions for each commit, try building for Windows (32 &
64 bit) and MS-DOS. Also, try to upload the resulting builds.

Note that the Linux version really should be built against bingcc and a
better SDL2 build (and maybe other variants like the different
renderers, etc, should be built), but this is worth a shot, at least.

There doesn't seem to be a good way of testing this other than
committing it, so let's see what happens!